Show new file as explicit entry in filer
This makes filer always shows an entry for the query itself in the file
list, which makes editing a new file more explicit.

Before, if the query didn't match anything "NO MATCHES FOUND" is shown
and if you press enter, clap opens a buffer named the current query.
It also wasn't possible to open a new buffer this way if the directory
was empty. A buffer with the name "Directory is empty" is opened no
matter what the input is.

Now, the current query is always shown as a file that can be opened,
which, at least for me, feels more intuitive and matches behavior of
similar interfaces (e.g. Helm find files in emacs).
This also works in empty directories. If you tab to an empty directory
and press enter with the empty query, the directory itself is opened.